categories

+++ Modules/Category/database/migrations/2024_04_18_184955_create_categories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->unsignedInteger('parent_id')->default(0);
+            $table->unsignedBigInteger('parent_id')->default(0);
             $table->timestamps();
         });
     }
